requirements.txt
kafka-python
nltk
psycopg2-binary
python-dotenv